Create FeatureDescriptor.py
Improved version of ConvFeatureDescriptor that runs a lot faster and only requires TensorFlow to be installed.
Leverages Keras' baked in pre-trained models to make predictions on either a single image or a batch of images. You can specify which specific "baked-in" architecture to use (they are all trained on ImageNet), or can define and specify your own pre-trained model (although it may need some re-tooling to work)

PIL is not required for the function to work, so can be refactored out later. It is just to demonstrate the test capabilities.

TLDR:
Takes in either a single image of shape (x,y,n_channels) or an array of images (number_images,x_dim,y_dim,n_channels), and produces a feature vector for each image which then can be processed by an arbitrary logits layer. This is useful for using classification models such as Random Forests or SVMs to classify image data, but can also be used with a conventional CNN with a fully connected layer on the features.